requirements.txt
GitPython==3.1.3
selenium==3.141.0
